adding add product functionality
