+++ src/hooks/useAddProduct.js
@@ -8,6 +8,11 @@ const ADD_PRODUCTS = gql`
 			description
 			price
 			image
+			ingredients
+			addOns
+			rating
+			calories
+			duration
 		}
 	}
 `;
